Confirmed by me

+++ MAGame/Assets/Scprits/Configs/PlayerWallet.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "PlayerWallet", menuName = "Configs/PlayerWallet")]
 public class PlayerWallet : ScriptableObject
 {
-    public int money;
+    public long money;
     public int moneyPerSecond;
 }
